+++ History.md
@@ -10,6 +10,7 @@
 * Bei Anträgen und Änderungsanträgen gibt es nun im Admin-Backend ein internes Feld "Interne Notiz", die nur Admins sehen können.
 * Das Beschlussdatum bei Gremien, die einen (Änderungs-)Antrag einreichen, wurde an vielen Stellen noch nicht angezeigt.
 * Eine Einstellungemöglichkeit, um das Login nur auf das Wurzelwerk zu beschränken (BDK)
+* Excel-Export der Anträge (BDK)
 
 2.3.0 - 2014-09-08
 ==================